renamed text input to cohortnameinput to be more descriptive
